Move db files to shared code
