+++ app/src/androidTest/java/com/omelan/cofi/ScreenshotCreator.kt
@@ -19,10 +19,10 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onRoot
 import androidx.compose.ui.test.performClick
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
-import com.omelan.cofi.model.PrepopulateData
 import com.omelan.cofi.pages.RecipeDetails
 import com.omelan.cofi.pages.RecipeEdit
 import com.omelan.cofi.pages.RecipeList
+import com.omelan.cofi.share.model.PrepopulateData
 import com.omelan.cofi.ui.CofiTheme
 import com.omelan.cofi.utils.ScreenshotsHelpers
 import org.junit.Rule
