+++ app/src/main/java/com/omelan/cofi/components/RecipeItem.kt
@@ -17,8 +17,8 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.omelan.cofi.R
-import com.omelan.cofi.model.Recipe
-import com.omelan.cofi.model.Step
+import com.omelan.cofi.share.Recipe
+import com.omelan.cofi.share.Step
 import com.omelan.cofi.share.StepType
 import com.omelan.cofi.ui.Spacing
 import com.omelan.cofi.ui.card
