+++ cofi-share/src/androidTest/java/com/omelan/cofi/share/AppDatabaseTest.kt
@@ -1,10 +1,9 @@
-package com.omelan.cofi.model
+package com.omelan.cofi.share
 
 import android.content.Context
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
-import com.omelan.cofi.share.StepType
 import com.omelan.cofi.utils.getOrAwaitValue
 import kotlinx.coroutines.runBlocking
 import org.junit.After
