+++ cofi-share/src/androidTest/java/com/omelan/cofi/share/MigrationsTest.kt
@@ -1,4 +1,4 @@
-package com.omelan.cofi.model
+package com.omelan.cofi.share
 
 import androidx.room.Room
 import androidx.room.testing.MigrationTestHelper
