+++ cofi-share/src/main/java/com/omelan/cofi/share/model/Step.kt
@@ -1,10 +1,49 @@
 package com.omelan.cofi.share
 
+import android.app.Application
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
+import androidx.annotation.WorkerThread
 import androidx.compose.ui.graphics.Color
+import androidx.lifecycle.AndroidViewModel
+import androidx.lifecycle.LiveData
+import androidx.room.*
+import com.omelan.cofi.share.model.AppDatabase
 import org.json.JSONArray
 import org.json.JSONObject
+
+@Dao
+interface StepDao {
+    @WorkerThread
+    @Query("SELECT * FROM step")
+    fun getAll(): LiveData<List<Step>>
+
+    @WorkerThread
+    @Query("SELECT * FROM step WHERE recipe_id IS :recipeId ORDER BY order_in_recipe ASC")
+    fun getStepsForRecipe(recipeId: Int): LiveData<List<Step>>
+
+    @WorkerThread
+    @Insert
+    suspend fun insertAll(vararg steps: Step)
+
+    @WorkerThread
+    @Insert
+    suspend fun insertAll(steps: List<Step>)
+
+    @WorkerThread
+    @Delete
+    suspend fun delete(step: Step)
+
+    @Query("DELETE FROM step WHERE recipe_id = :recipeId")
+    suspend fun deleteAllStepsForRecipe(recipeId: Int)
+}
+
+class StepsViewModel(application: Application) : AndroidViewModel(application) {
+    private val db = AppDatabase.getInstance(application)
+
+    fun getAllStepsForRecipe(recipeId: Int) = db.stepDao().getStepsForRecipe(recipeId)
+    fun getAllSteps() = db.stepDao().getAll()
+}
 
 enum class StepType(
     val color: Color,
@@ -42,10 +81,12 @@ enum class StepType(
 }
 
 open class StepTypeConverter {
+    @TypeConverter
     open fun stepTypeToString(type: StepType): String {
         return type.name
     }
 
+    @TypeConverter
     open fun stringToStepType(type: String): StepType {
         return when (type) {
             StepType.ADD_COFFEE.name -> StepType.ADD_COFFEE
@@ -57,14 +98,15 @@ open class StepTypeConverter {
     }
 }
 
-open class StepShared(
-    open val id: Int = 0,
-    open val recipeId: Int = 0,
-    open val orderInRecipe: Int? = null,
-    open val name: String,
-    open val time: Int? = null,
-    open val type: StepType,
-    open val value: Int? = null,
+@Entity
+data class Step(
+    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @ColumnInfo(name = "recipe_id") val recipeId: Int = 0,
+    @ColumnInfo(name = "order_in_recipe") val orderInRecipe: Int? = null,
+    val name: String,
+    val time: Int? = null,
+    val type: StepType,
+    val value: Int? = null,
 )
 
 private const val jsonName = "name"
@@ -73,7 +115,7 @@ private const val jsonType = "type"
 private const val jsonOrderInRecipe = "orderInRecipe"
 private const val jsonValue = "value"
 
-fun StepShared.serialize(): JSONObject = JSONObject().let {
+fun Step.serialize(): JSONObject = JSONObject().let {
     it.put(jsonName, name)
     it.put(jsonTime, time)
     it.put(jsonValue, value)
@@ -82,7 +124,7 @@ fun StepShared.serialize(): JSONObject = JSONObject().let {
     it
 }
 
-fun List<StepShared>.serialize() = JSONArray().let {
+fun List<Step>.serialize() = JSONArray().let {
     forEach { step -> it.put(step.serialize()) }
     it
 }
@@ -93,7 +135,7 @@ fun JSONObject.getIntOrNull(key: String) = try {
     null
 }
 
-fun JSONObject.toStep(recipeId: Long = 0) = StepShared(
+fun JSONObject.toStep(recipeId: Long = 0) = Step(
     name = getString(jsonName),
     recipeId = recipeId.toInt(),
     time = getIntOrNull(jsonTime),
@@ -102,8 +144,8 @@ fun JSONObject.toStep(recipeId: Long = 0) = StepShared(
     type = StepTypeConverter().stringToStepType(getString(jsonType)),
 )
 
-fun JSONArray.toSteps(recipeId: Long = 0): List<StepShared> {
-    var steps = listOf<StepShared>()
+fun JSONArray.toSteps(recipeId: Long = 0): List<Step> {
+    var steps = listOf<Step>()
     for (i in 0 until length()) {
         steps = steps.plus(getJSONObject(i).toStep(recipeId))
     }
